#209 Extract report class in multiple classes
