+++ api/gitlab/report/report_json.py
@@ -1,0 +1,48 @@
+report_dict = {
+    "branches": {
+      "name": 0
+    },
+    "commits": {
+        "last_commit": {
+            "title": 0,
+            "author_name": 0,
+            "author_email": 0,
+            "authored_date": 0
+        },
+        "number_of_commits": 0
+    },
+    "project": {
+        "name": 0,
+        "web_url": 0
+    },
+    "pipelines": {
+        "number_of_pipelines": 0,
+        "succeded_pipelines": 0,
+        "failed_pipelines": 0,
+        "percent_succeded": 0,
+        "current_pipeline_id": 0,
+        "current_pipeline_name": 0,
+        "recents_pipelines": {
+            "last_7_days": {
+                "number_of_pipelines": 0,
+                "percent_succeded": 0,
+                "percent_failed": 0,
+                "succeded_pipelines": 0,
+                "failed_pipelines": 0
+            },
+            "last_30_days": {
+                "number_of_pipelines": 0,
+                "percent_succeded": 0,
+                "percent_failed": 0,
+                "succeded_pipelines": 0,
+                "failed_pipelines": 0
+            }
+          }
+    },
+    "pipelines_times": {
+        "average": 0,
+        "lower": 0,
+        "higher": 0,
+        "total": 0
+    }
+}
